requirements.txt
numpy
tiktoken
scikit-learn
sentence-transformers
faiss-cpu
fastapi
uvicorn
